Updated server models
